removes loading spinner for expand card, looked weird
