+++ src/components/Dashboard/PageCard/PageCard.sass
@@ -12,9 +12,9 @@
     -webkit-user-select: none
 
 
-.pageCard_header:hover
-    opacity: .8
-    background-color: lightgrey
+// .pageCard_header:hover
+//     opacity: .8
+//     background-color: lightgrey
 
 
 .pageCard_title
